Move triggerable to own file
