+++ src/collision_handlers/polling_collision_handler.py
@@ -11,7 +11,7 @@ import src.game_object.bomb as bomb
 import src.game_object.tile as tile
 import src.graphics as graphics
 from src.game_object.pressure_plate import PressurePlate
-from src.game_object.tile import Triggerable
+from src.game_object.triggerable import Triggerable
 
 
 class PollingCollisionHandler():
